agentic-hackathon/requirements.txt
openai>=1.40.0
langchain>=0.2.10
langchain-openai>=0.1.22
langchain-community>=0.2.9
langchain-text-splitters>=0.2.2
langgraph>=0.2.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.0
